Updated futures with more simplicity
